Conv2Conv: move CNNDecoder into onmt/modules/Conv2Conv.py
..., while at it, collapse the ConvDecoder into CNNDeocder. It is
not necessary to add another wrapper.
