+++ onmt/modules/Conv2Conv.py
@@ -4,16 +4,20 @@ Implementation of "Convolutional Sequence to Sequence Learning"
 import torch
 import torch.nn as nn
 import torch.nn.init as init
-import onmt.modules
-from onmt.modules.WeightNorm import WeightNormConv2d
 import torch.nn.functional as F
 from torch.autograd import Variable
 
-scale_weight = 0.5 ** 0.5
+import onmt.modules
+from onmt.modules.WeightNorm import WeightNormConv2d
+from onmt.Translator import DecoderState
+from onmt.modules import aeq
+
+
+SCALE_WEIGHT = 0.5 ** 0.5
 
 
 def shape_transform(x):
-    # tranform the size of the tensor to fit for conv input
+    """ Tranform the size of the tensors to fit for conv input. """
     return torch.unsqueeze(torch.transpose(x, 1, 2), 3)
 
 
@@ -48,7 +52,7 @@ class StackedCNN(nn.Module):
     def forward(self, x, hidden=None):
         for conv in self.layers:
             x = x + conv(x)
-            x *= scale_weight
+            x *= SCALE_WEIGHT
         return x
 
 
@@ -88,29 +92,28 @@ class ConvEncoder(nn.Module):
         return outputs.squeeze(3), emb_remap.squeeze(3)
 
 
-class ConvDecoder(nn.Module):
-
-    def __init__(self, input_size, hidden_size, num_layers, dropout,
-                 cnn_kernel_width):
+class CNNDecoder(nn.Module):
+    """
+    Decoder built on CNN, which consists of resduial convolutional layers,
+    with ConvMultiStepAttention.
+    """
+    def __init__(self, num_layers, hidden_size, attn_type,
+                 copy_attn, cnn_kernel_width, dropout, embeddings):
         """
-          Conv Decoder consists of layers of resduial conv layer
-          and ConvMultiStepAttention. encoder the sequence of target token.
-          Args:
-                input_size: dim of target token vector.
-                hidden_dim: the size of channel in conv.
-                num_layers: the num of conv layer.
-                dropout: dropout rate.
-                cnn_kernel_width: the width of the kernel in cnn.
+        See make_decoder() comment for arguments description.
         """
+        super(CNNDecoder, self).__init__()
 
-        super(ConvDecoder, self).__init__()
-
+        # Basic attributes.
+        self.decoder_type = 'cnn'
         self.num_layers = num_layers
-        self.input_size = input_size
-        self.cnn_kernel_width = cnn_kernel_width
         self.hidden_size = hidden_size
+        self.cnn_kernel_width = cnn_kernel_width
+        self.embeddings = embeddings
         self.dropout = dropout
 
+        # Build the CNN.
+        input_size = self.embeddings.embedding_dim
         self.linear = nn.Linear(input_size, self.hidden_size)
         self.conv_layers = nn.ModuleList()
         for i in range(self.num_layers):
@@ -123,20 +126,63 @@ class ConvDecoder(nn.Module):
             self.attn_layers.append(
                 onmt.modules.ConvMultiStepAttention(self.hidden_size))
 
-    def forward(self, target_emb, encoder_out_top, encoder_out_combine):
+        # CNNDecoder has its own attention mechanism.
+        # Set up a separated copy attention layer, if needed.
+        self._copy = False
+        if copy_attn:
+            self.copy_attn = onmt.modules.GlobalAttention(
+                hidden_size, attn_type=attn_type)
+            self._copy = True
+
+    def forward(self, input, context, state):
         """
-        decode current timestep token by history target token with
-        ource encoder vector.
+        Forward through the CNNDecoder.
         Args:
-            target_emb: the embedding vector of target sequence
-            encoder_out_top: the top output of source sequence
-            through ConvEncoder
-            encoder_out_combine: encoder_out_top + source_emb.
+            input (LongTensor): a sequence of input tokens tensors
+                                of size (len x batch x nfeats).
+            context (FloatTensor): output(tensor sequence) from the Encoder
+                        CNN of size (src_len x batch x hidden_size).
+            state (FloatTensor): hidden state from the Encoder CNN for
+                                 initializing the decoder.
+        Returns:
+            outputs (FloatTensor): a Tensor sequence of output from the Decoder
+                                   of shape (len x batch x hidden_size).
+            state (FloatTensor): final hidden state from the Decoder.
+            attns (dict of (str, FloatTensor)): a dictionary of different
+                                type of attention Tensor from the Decoder
+                                of shape (src_len x batch).
         """
-        emb_reshape = target_emb.contiguous().view(
-            target_emb.size(0) * target_emb.size(1), -1)
+        # CHECKS
+        assert isinstance(state, CNNDecoderState)
+        input_len, input_batch, _ = input.size()
+        contxt_len, contxt_batch, _ = context.size()
+        aeq(input_batch, contxt_batch)
+        # END CHECKS
+
+        if state.previous_input is not None:
+            input = torch.cat([state.previous_input, input], 0)
+
+        # Initialize return variables.
+        outputs = []
+        attns = {"std": []}
+        assert not self._copy, "Copy mechanism not yet tested in conv2conv"
+        if self._copy:
+            attns["copy"] = []
+
+        emb = self.embeddings(input)
+        assert emb.dim() == 3  # len x batch x embedding_dim
+
+        tgt_emb = emb.transpose(0, 1).contiguous()
+        # The output of ConvEncoder.
+        src_context_t = context.transpose(0, 1).contiguous()
+        # The combination of output of ConvEncoder and source embeddings.
+        src_context_c = state.init_src.transpose(0, 1).contiguous()
+
+        # Run the forward pass of the CNNDecoder.
+        emb_reshape = tgt_emb.contiguous().view(
+            tgt_emb.size(0) * tgt_emb.size(1), -1)
         linear_out = self.linear(emb_reshape)
-        x = linear_out.view(target_emb.size(0), target_emb.size(1), -1)
+        x = linear_out.view(tgt_emb.size(0), tgt_emb.size(1), -1)
         x = shape_transform(x)
 
         pad = Variable(torch.zeros(x.size(0), x.size(1),
@@ -148,6 +194,43 @@ class ConvDecoder(nn.Module):
             new_target_input = torch.cat([pad, x], 2)
             out = conv(new_target_input)
             c, attn = attention(base_target_emb, out,
-                                encoder_out_top, encoder_out_combine)
-            x = (x + (c + out) * scale_weight) * scale_weight
-        return x.squeeze(3).transpose(1, 2), attn
+                                src_context_t, src_context_c)
+            x = (x + (c + out) * SCALE_WEIGHT) * SCALE_WEIGHT
+        output = x.squeeze(3).transpose(1, 2)
+
+        # Process the result and update the attentions.
+        outputs = output.transpose(0, 1).contiguous()
+        if state.previous_input is not None:
+            outputs = outputs[state.previous_input.size(0):]
+            attn = attn[:, state.previous_input.size(0):].squeeze()
+            attn = torch.stack([attn])
+        attns["std"] = attn
+        if self._copy:
+            attns["copy"] = attn
+
+        # Update the CNNDecoderState.
+        state.resetPrevious(input)
+
+        return outputs, state, attns
+
+
+class CNNDecoderState(DecoderState):
+    def __init__(self, context, enc_hidden, input=None):
+        self.init_src = (context + enc_hidden) * SCALE_WEIGHT
+        self.previous_input = input
+        self.all = (self.previous_input,)
+
+    def _resetAll(self, all):
+        vars = [(Variable(a.data if isinstance(a, Variable) else a,
+                          volatile=True))
+                for a in all]
+        self.previous_input = vars[0]
+        self.all = (self.previous_input,)
+
+    def repeatBeam_(self, beamSize):
+        self.init_src = Variable(
+            self.init_src.data.repeat(1, beamSize, 1), volatile=True)
+
+    def resetPrevious(self, input):
+        self.previous_input = input
+        self.all = (self.previous_input,)
